block_synchronizer/tests: move tests to separate files
Move global state synchronizer and trie accumulator tests to separate
files.
